protected over private / logic change in getDeviceInfo
Replaced all instances of private with protected. This will allow others to extend the Nest class into subclasses where functionality may not be for the good of the group, but for their own private use.

With the implementation of Auto / Manual Eco Mode versus Auto/Manual Away the logic for what the thermostat is actually doing has changed. My new logic accounts for scenarios where the structure is still away - but the house has already switched itself back to schedule.  Manual_away really shouldn't be used anymore because it doesn't exist as an option. Although, you can change the structure to away via app/web manual - it should no longer be referred to as "manual away" as it doesn't actually trigger going into manual-eco, but rather forces thermostat into auto-eco and is more for testing mode of cameras.